Move selector according to new structure
