+++ src/pages/search/layout/selector/PolygonIcon.tsx
