+++ src/pages/search/layout/selector/RemoveIcon.tsx
