+++ src/pages/search/layout/selector/SearchAreas.tsx
